08/python_project/requirements.txt
Flask
pandas
pyjwt